devtools files are ts files
Unused files are deleted.
