--- packages/devtools/bin/deploy-contracts.js
@@ -1,6 +1,0 @@
-#! /usr/local/bin/node
-
-"use strict";
-const { deployContracts } = require("../utils/deployContracts");
-
-deployContracts();
