--- packages/devtools/bin/start-ganache.js
@@ -1,7 +1,0 @@
-#! /usr/local/bin/node
-
-"use strict";
-let argv = require("yargs").argv;
-const { startGanache } = require("../utils/startGanache");
-
-startGanache(argv);
