+++ packages/devtools/src/utils/expect-revert.ts
@@ -1,0 +1,3 @@
+export async function expectRevert(fn, pattern?: string | RegExp) {
+  await expect(fn()).rejects.toThrowError(pattern);
+}
