--- packages/devtools/utils/artifactParsers.js
@@ -1,24 +1,0 @@
-module.exports = {
-  parseContractAddress: function(pathToArtifact, networkId) {
-    const contractArtifact = require(pathToArtifact);
-    const network = contractArtifact.networks[networkId];
-    if (!network) {
-      console.error(
-        `There is no network information for network id ${networkId} in the artifact file ${pathToArtifact}`
-      );
-      process.exit(1);
-    }
-    return network.address;
-  },
-  parseABI: function(pathToArtifact) {
-    const contractArtifact = require(pathToArtifact);
-    const abi = contractArtifact.abi;
-    if (!abi) {
-      console.error(
-        `There is no abi information in the artifact file ${pathToArtifact}`
-      );
-      process.exit(1);
-    }
-    return JSON.stringify(abi);
-  }
-};
