--- packages/devtools/utils/delay.js
@@ -1,7 +1,0 @@
-module.exports = {
-  delay: function(ms = 200) {
-    return new Promise(resolve => {
-      setTimeout(resolve, ms);
-    });
-  }
-};
