--- packages/devtools/utils/expectRevert.js
@@ -1,5 +1,0 @@
-module.exports = {
-  expectRevert: async (fn, pattern) => {
-    await expect(fn()).rejects.toThrowError(pattern);
-  }
-};
